gpu-server/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.28.0
transformers>=4.40.0
accelerate>=0.30.0
safetensors>=0.4.0
controlnet-aux>=0.0.9
opencv-python>=4.9.0
Pillow>=10.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.9
simple-lama-inpainting>=0.1.0
